:truck: chore(api) move errors helper to src/
