+++ api/src/prescription/campaign-participation/domain/usecases/send-shared-participation-results-to-pole-emploi.js
@@ -1,4 +1,4 @@
-import * as httpErrorsHelper from '../../../../../lib/infrastructure/http/errors-helper.js';
+import * as httpErrorsHelper from '../../../../../src/prescription/campaign-participation/infrastructure/errors-helper.js';
 import { httpAgent } from '../../../../shared/infrastructure/http-agent.js';
 import { logger } from '../../../../shared/infrastructure/utils/logger.js';
 import { PoleEmploiPayload } from '../../infrastructure/externals/pole-emploi/PoleEmploiPayload.js';
