+++ api/src/prescription/campaign-participation/infrastructure/errors-helper.js
