+++ api/tests/prescription/campaign-participation/integration/infrastructure/errors-helper_test.js
@@ -1,5 +1,5 @@
-import { serializeHttpErrorResponse } from '../../../../lib/infrastructure/http/errors-helper.js';
-import { expect } from '../../../test-helper.js';
+import { serializeHttpErrorResponse } from '../../../../../src/prescription/campaign-participation/infrastructure/errors-helper.js';
+import { expect } from '../../../../test-helper.js';
 
 describe('serializeHttpErrorResponse', function () {
   describe('when http error data is null', function () {
